fix(review): apply PR #161 self-review HIGH findings
修正 code-reviewer 在 PR #161 self-review 中的 2 個 HIGH：

1. saveButtonLabel module coupling: 把 pure helper + UploadProgress type
   抽到 src/lib/save-button-label.ts，讓單元測試不必再 mock 25 行
   Firebase modules（expense-form.tsx 內的 Firebase imports 變動
   不會再意外 break 此 helper 的測試）

2. aria-live 誤用在 <button> 元素上：NVDA/VoiceOver 對 interactive
   widget 上的 live 屬性支援不一致，改為 sibling visually-hidden
   <div role="status" aria-live="polite" aria-atomic="true">
   在儲存期間承載進度文字，按鈕本身只保留可見 label

附帶 UX 改進：
- 新增「準備上傳 N 張...」狀態 → 取代 current=0 時的「上傳中 0/N 張」
  減少使用者認知負擔（「0 張」視覺上看起來矛盾）
- 測試新增此 case 覆蓋

MEDIUM（unmounted state update）暫不修：React 18 production 靜默
略過，只有 dev warning；isMounted ref 成本不划算。
